Bugfix one-hot encoding of targets (#50)
The targets for the UNet baseline became `(B, 1, D, H, W, C)` after the one-hot encoding, which includes a dummy dimension that should not be there. Subsequently, the `moveaxes` is incorrect. This fixes the one-hot encoding and performs this outside the loss function, to enable easier model development with e.g. different loss functions.